Loading Screen Complete
Uses Javascript to listen for a submit.

On Submit,
If all the fields are correct then we can set display = none for the div wit id "donation-page" that contains the main form, making it disappear, and make the "loading-element" div appear by setting its display to flex.

We also add a submit event listener to prevent the default submit behavior from happening so that the loading screen can trigger.